[code move] move pp into a separate mixin (#11838)
